fix: gradio demo parameters
